13: Fixed review findings.
