+++ pom.xml
@@ -4,7 +4,7 @@
     <groupId>com.exasol</groupId>
     <artifactId>exasol-virtual-schema-lua</artifactId>
     <version>0.2.0</version>
-    <name>Exasol Row Level Security (Lua)</name>
+    <name>Exasol Virtual Schema (Lua)</name>
     <description>This projects provides a plug-in to the Exasol database that adds per-row access control.</description>
     <url>https://github.com/exasol/exasol-virtual-schema-lua/</url>
     <properties>
